Moved 3d audio test scripts to utils.
